+++ otherfile/profiles/ws_profile.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+"""Profile file."""
+from pyrgg import *
+import pyrgg.engines.watts_strogatz as ws_engine
+import random
+
+os.environ["PYRGG_TEST_MODE"] = "1"
+random.seed(400)
+
+ws_engine.gen_using(
+    dimacs_maker,
+    'profile',
+    {
+        'vertices': 10,
+        'mean_degree': 2,
+        'rewiring_probability': 0.3,
+    }
+)