+++ pyrgg/params.py
@@ -16,6 +16,7 @@ MENU_ITEMS = {
         3- Erdos-Renyi - G(n, m)
         4- Stochastic Block Model - G(C, P)
         5- Barabasi-Albert - G(n, k)
+        6- Watts-Strogatz - G(n, k, p)
         """
     )],
     2: ["file_name", "- File Name (Not Empty) : "],
@@ -69,6 +70,7 @@ ENGINE_MENU = {
     3: "er",
     4: "sbm",
     5: "ba",
+    6: "ws",
 }
 
 ENGINE_MENU_INV = {v: k for k, v in ENGINE_MENU.items()}
@@ -112,12 +114,19 @@ BA_ENGINE_PARAMS = {
     2: ["attaching_edge_number", "- Number of Edges to Attach to a New Node (0 < k < n) : "],
 }
 
+WS_ENGINE_PARAMS = {
+    1: ["vertices", "- Vertices Number (n >= 0) : "],
+    2: ["mean_degree", "- Mean Degree (should be a positive even number) : "],
+    3: ["rewiring_probability", "- Rewiring Probability (0 <= p <= 1) : "],
+}
+
 ENGINE_PARAM_MAP = {
     1: PYRGG_ENGINE_PARAMS,
     2: ERG_ENGINE_PARAMS,
     3: ER_ENGINE_PARAMS,
     4: SBM_ENGINE_PARAMS,
     5: BA_ENGINE_PARAMS,
+    6: WS_ENGINE_PARAMS,
 }
 
 OUTPUT_FORMAT = {i: output_format[1:].upper()